requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
facenet-pytorch==2.5.2
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
Pillow==8.3.2
requests==2.26.0
# torch==1.9.0
# torchvision==0.10.0
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2

# --find-links https://download.pytorch.org/whl/torch_stable.html
# torch==1.9.0+cpu
gunicorn==20.0.4

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
torchvision==0.10.0+cpu


